requirements.txt
py-cord
python-dotenv
google-genai
better-profanity